Skip serializing ResolutionResult if requested
The StreamingResolutionResultBuilder visits a resolved graph, captures any data needed to produce a ResolutionResult, and serializes it.
Later on, if the user requests a ResolutionResult, we de-serialize and cache the result.
We do this, because in most cases the ResultionResult is not used by the user.

However, in some cases, before resolution, we know that the user requested a ResolutionResult.
In these cases, we can skip the serialization round-trip time and load the result directly into memory.