Implement ParallelListener interface
Introduce a new ParallelListener interface, marking listners that can accept events
from multiple threads. This allows us to avoid locks in the event dispatch path.

This annotation should be used on listeners that are called very often from multiple threads.
We replace implementations of `ProjectComponentObservationListener` with implementations that
are parallel-safe and marked with this annotation.

We also modify the semantics of `@StatefulListener` to forbid _removing_ listeners as well as adding
them, after events have been emitted. This is because we achieve lock-free thread-safety for stateful
listeners by marking these listeners as immutable after events have been emitted for those listeners