Avoid flakiness in SupportedBuildJvmVersionIntegrationTest
Make sure to write the raw vendor in case we don't recognise the vendor. Previously if we found a JVM with an unknown vendor,
we would write UNKNOWN as the vendor, and then fail when checking if that JVM was compatible.

Also require isolated daemons for SupportedBuildJvmVersionIntegrationTest since before it was possible that the test picked up
a compatible daemon from another test but with a different java home. We need are checking that the java home configured via
installation paths is the same as the java home used by the daemon. It is a bit weird that the launcher will pick a running daemon
started with a jvm that is not in its list of discovered JVMs, but as long as that happens, we need isolated daemons