Avoid backreference from variant to component
This backreference was there pretty much only to support calculating the public view of a variant when serializing
the variant when assembling the serialized resolution result. We reoganize the way serialization works by separating
adhoc and non-adhoc serialization. This allows us to much more easily track the component of the variants that
are being serialized, so that we do not need the variant to hold a reference to the owning component.

Backreferences like these makes it very annoying to efficiently and cleanly manage mutable state. In this specific case,
in order to instantiate a variant, I must have an already instantiated component instance. Future work that aims to
vastly clean up the way local component state is created would rely on creating component state instances that
delegate to another, but add additional variants. Without this change, further enhancements to local component state
would not be feasible.