Polish domain object collections
Delete PendingSource, moving all methods to ElementSource. No other class/interface extended from PendingSource other than ElementSource.
Add some javadoc to ElementSource

Delete WithEstimatedSize and WithMutationGuard. Usage of these methods required us to call instanceof using helpers like Estimates.estimatedSizeOf and MutationGuards.of
Helper methods like those are impresise, as we are unsure what sort of object we are calling the helper on.
Furthermore, there are a number of classes that implemented these interface but never had the helper methods called on them.

Instead, we move getMutationGuard and estimatedSize to a new interface, DomainObjectCollectionInternal, which is implemented by DefaultDomainObjectCollection and DelegatingDomainObjectCollection

We merge AbstractMuationGuard and DefaultMutationGuard, as the default implementation was the sole implementation, and therefore there was no reason to have the logic split.