Initialize report containers upon construction
Report containers are immutable, but contain an initial set of elements. Subclasses of DefaultReportContainer
currently call the super constructor, then add their own elements in their own constructor. In order to
enforce immutability, DefaultReportContainer overrides teh assertMutableCollectionContents, throwing an
exception unconditionally.

In order for subclasses to avoid this exception, the add method is overridden to not trigger the exception,
and subclasses add their elements in their constructor. This pattern is unconventional among domain object containers.

In this commit, we rework the way these containers initialize their elements, so that they pass their initial elements to
the DefaultReportContainer constructor. This way, after the superconstructor returns, the container is fully immutable.
Future work will remove the need for the 'initialized' variable and 'assertMutableCollectionContents' method, but for
now these mechanisms are used to only turn on immutability after the container sets its initial elements.

Further work was done to remove the requirement for a report to keep a reference to the task that generates that report,
as the Task referene was only used to generate a description for the report.