Require MutationGuard to be set explicitly for Collections
Each collection has a MutationGuard embedded in its ElementSource. There is also a project-wide MutationGuard owned by the CrossProjectConfigurator.
This project-wide guard is additionally triggered when afterEvaluate/beforeEvaluate execute, preventing (currently some) containers from registering
lazy actions and mutating containers within these callbacks.
However, not all containers are properly initialized with this 'parentMutationValidator'. Many containers did not have this validator and defaulted
to the MutationValidators.identity() implementation. As a result, many containers _could_ be mutated within afterEvaluate/beforeEvaluate.

This PR updates the constructors of the various container/collection implementations to require a MutationGuard to be set, removing the constructors
that default to MutationGuards.identity(). All containers that previousuly defaulted to the identity guard explicitly set that as their guard now,
as properly configuring them to the project-wide guard would be a breaking change.

Future work should add infrastructure to allow us to migrate all containers away from MutationGuards.identity() by adding a deprecation cycle.
This change is out of scope for this commit, which aims to clean up and properly document the containers that do use the idenity guard.

This commit makes additional changes in the same area:
- Consolidate constructors of all containers/collections, as many implementations had a number of constructors which became confusing
- Add javadoc
- Migrate some containers which previously required a custom Namer to use Named.namer.forType(), and add javadoc to prefer types that implement Named
- SourceSet, Report, and JavaToolchainRepository all now implement Named.