Clarify the purpose of MutationGuard
MutationGuard is poorly named. In practice, it tracks whether the current thread is operating a lazy action
on a given domain object or set of domain objects. Documentation and method names were updated to reflect this.

MutationGuard should also be renamed, but given that we have other changes to make to where it is used and due
to the number of places it is used, we are leaving it with its current name for now.