Always cache Attribute compatibility and disambiguation rules
Compatibility rules especially are called in tight loops in the middle of dependency resolution.
We have seen builds that rely on custom rules heavily spend a huge amount of time during resolution
instantiating these rules over and over.

This commit updates the rule instantiator to cache the value for the rule so that we do not waste time
instantiating the rule over and over again.

We expect these rules to be deterministic. Both AGP and KGP rules were verified to be deterministic and
not have side-effects in their constructors.