Remove transform dependencies resolver factory
This type, TransformUpstreamDependenciesResolverFactory, hid a very messy state cycle that
currently backs the resolution of artifact transform dependencies.

In short, the @InputArtifactDependencies annotation for artifact transforms allow a transform
to consume its dependency artifact. We leverage the resolved graph to find the dependenices that
are contained within the subgraph that is a child of the node we are resolving. We then resolve
artifacts for that subgraph using the input attributes for the transform we are executing.

This means that in order to resolve artifacts, we must be able to resolve artifacts. This introduces
a difficult logic cycle in the code, which is currently backed by a configuration. The artifact selection
process takes in the lazy results of resolution. It is expected to not call these methods until resolution
is performed. These results are backed by a cache in the configuration that ensures the currently-under-construction
results are passed back to itself when the artifact selection results queries these resolution results.

This pattern makes it incredibly complicated to implement artifact selection, as it relies on this logic loop/cache.

This is the initial step to simplifying the logic, which is to expose the actual data from the configuration that
TransformUpstreamDependenciesResolverFactory was hiding. Future work should be done to "push back" the time that
these results are required, so that we pass them in during artifact selection (or resolution). The later we can provide
these results, the less we implicitly rely on the cache that lives in DefaultConfiguration.