Deprecate manually adding configurations to configuration container
Detached configurations and copied configurations, which are also detached configurations, should not be added to the configuration container.
As such, we deprecate calling any add method to the container, restricting the user to only add configurations using the standard factory methods.

Furthermore, we modify the roles given to copied configurations. Previously, we used the roles of the original configuration.
However, when a configuration is copied, all of its dependencies, even from the configurations it extends from, are accumulated and added to the detached configuration.
This means that the copied configuration will have dependencies and will be resolvable.
As such, it must be a resolvable dependency scope.