Avoid Mockk in build-logic tests
While upgrading to Java 21, mockk was having trouble mocking static methods.
Bumping the mockk version did not solve the issue, and it was easier to get rid of mock rather
than determine the root cause.

Instead, we update DetermineBaselines to accept a CommandExecutor and provide a test implementation during testing.