Deprecate ResolvedConfiguration and LenientConfiguration getFiles
The getFiles method on these two classes are easily replaced by existing methods. The deprecated methods
are not CC compatible and do not carry task dependencies, like their replacements do.

Furthermore, we improve documentation for these legacy types, ResolvedConfiguration, LenientConfiguration,
and ArtifactResolutionQuery. We add javadoc that states these types are legacy, should not be used, and
will be removed in future versions. Additionally, we try to push users away from some ResolvedConfiguration
and LenientConfiguration methods towards ones that will not be deprecated in the future, even though
we do not deprecate them now.

Finally, we add some unrelated tests to verify the behavior of artifact view configured to resolve leniently.