Pull out DefaultAttributeSelectionSchema
This class was embedded in DefaultAttributesSchema but many tests attempted to directly test ti
We pull it out so its can be tested separately.

Then, we update a bunch of tests to reflect this change.
Furthermore, we refactor some tests to prepare them for immutable attribute schemas by using
factory methods accepting a closure in order to configure schemas

Also, we merge AttributeSelectionUtils functionality into DefaultAttributeSelectionSchema
as we no longer need that logic separated for testing.

Finally, we create a new 'matching' package to contain attribute matching related classes
and move some existing classes there