Introduce ImmutableAttributeSchema
Introduce ImmutableAttributeSchema, which can be compared with equality. We introduce a factory
for these schemas which intern instances. Then, we introduce another factory for attribute matchers
keyed on the input schema. This allows us to 'hoist' up various attribute matching caches
to the build session scope.

There is still much work to be done. Artifact transform selection still requires a similar hoist.
Plus, there is de-duplication to be done with the artifact type registry