Switch order of producer and consumer when merging strategies
The order of the attributesByName field in an ImmutableAttributesSchema is derived from the order of the keys
in the strategies map. Keys that appear later that have the same name as keys that appear earlier will take
precedence. We want consumer keys to take precedence, due to backward compatibility behavior, so we put
producer strategies in the list first.

The order here really should not matter, but due to the way things used to work, the getAttributeByName method
of an attributes schema should return the consumer attribute with that name instead of the producer attribute
with that name. The root problem here is that the consumer and producer are able to disagree here, or rather
that we even feel the need to request attributes by name at all.

The only place we try to get an attribute from a schema by name is the tryRehydrate method, where we
attempt to upgrade a desugared attribute to the proper attribute as known by the schema. We then use this key
to fetch the compatibility and disambiguation rules for attribute matching. If two schemas were merged and
had the 'same' attributes with different Class definitions, perhaps the same attribute but loaded in different
classloaders, we will only get the registered rules for the consumer.

This problem is outlined by the new test added, that verifies when we load the same attribute in two different
classloaders, the producer rules are ignored.