When running performance tests, no daemon should mean "cold daemon"
The Gradle profiler distinguishes between a "no daemon" and a
"cold daemon".

A cold daemon is where a new daemon is started every time a measurement is performed.
A no daemon is where we attempt to run the build in the launcher.

We will eventually deprecate/remove support for running inside the launcher,
so this isn't the scenario we want to test.

We have gone through some hoops to try to make the launcher compatible to
run builds inside it, but this is also not something real world builds will
(or can) do in some cases.