Reintroduce capability conflict bugfixes
When we try to determine if a conflict is registered, it is possible that we
register the conflict with a node that is currently not selected due to a version conflict,
while its replacement is still in the queue and not yet registered with the capability
conflict handler.

At the time of deselecting modules from the conflict with the non-selected candidate, we
also deselect the replaced version node that is curerntly in the queue. When we process that node,
we do not register a capability conflict for it since it is now deselected.

The fix is to be sure that we do not register conflicts with nodes that are deselected