Rewrite capability conflict detection
The process of detecting capabiltity conflicts was previously very confusing. The logic
was separated among several files and the reasoning for various edge cases were not
well explained.

This commit migrates the majority of the logic into DefaultCapabilitiesConflictHandler.
The logic is simplified, as now we keep track of nodes as they enter and leave the graph.
Implicit and explicit capabilities for nodes are tracked by the conflict handler, so we
no longer need to filter nodes from conflicts when they are resolved, as capabilities
currently in the graph are constantly updated as nodes enter and leave the graph.

Further, we no longer need to iterate over every module of the graph each time we
see a new implicit capability, as we already know which nodes contribute which capabilities.
Keeping track of implicit capabilities here is much more sensible than iterating over the
entire graph for each implicit capability that may be in conflict.