Resolve substituted project dependencies when performing dependency mapping
If attempting to resolve a module dependency substituted for a project dependency using variant
coordinate mapping, we would fail since the project dependency resolver was not cross-build.
We update the project dependency resolver to be a build tree scoped service so that it can
resolve publications from all projects in the build tree and not just the current build.

Furthermore, we failed to resolve these substituted dependencies when using component
dependency mapping, as the component coordinate mapping implementation only looked at
resolved local components when resolving module dependencies. A proper implementation
would look at requested vs selected components, so we update the existing variant
mapping implementation that performs this logic to also implement the component mapping interface.

While updating the project dependency resolver, we also update the project publication registry
to not hold onto Project instances but instead operate solely on ProjectIdentity references.