Fix cache entry retention with configuration cache
Previously, we were immediately converting the `removeUnusedEntriesAfter` setting
to an absolute timestamp. This timestamp was then persisted in the configuration cache.

With this change, we retain the `removeUnusedEntriesAfter` as an interval in ms,
and only convert to an absolute timestamp when constructing the cache instance
This means that the interval is always based on the current build execution, rather
than the time when the configuration was persisted for CC.

Fixes #30467