Allow setting an absolute timestamp fo cache-cleanup
For certain cases, it's desirable for the cache entry retention to be specified
as an absolute timestamp, rather than an interval in days. One use case is to force
removal of unused entries in the `setup-gradle` GitHub Action.

For this purpose, an new `entryRetentionTimestamp` property is introduced, that contains an
absolute timestamp. The existing `entryRetentionMillis` property contains an interval
and has been renamed to `entryRetentionAge`.

Both of these properties are configurable, and both are retained in configuration-cache state.
If `entryRetentionTimestamp` is set, then the value for `entryRetentionAge` will be ignored.