Slim down VariantResolvingArtifactset and cache the implicit instance for a given node at the build-tree level
These types were a major contributor to retained heap of a configuration after resolution. We remove as many
fields as we can from it by passing any services in to its select method instead of storing them as a field.

Further instead of creating an instance of these for each node in the graph that contributes artifacts to the result, we
cache the 'implicit' artifact set for a given node, which is the artifact set that is not modified by a dependency.

These implicit artifact sets are by far the most common, as most dependencies do not modify artifacts (they do not have attributes, capabilities, excludes, or requested artifacts).
This is only done for project components, as project components have a single instance ID per build tree. We still
need some deduplication for external components if we want this new cache to be effective for them.