Slim down dependencies
Builds create a lot of dependencies. These dependenices are kept in memory for the entirety of a build when configuration
cache is not enabled. Or, during sync, they are kept in memory for the entirety of sync as sync does not use CC.

Many dependencies do not have artifacts, excludes, or rejected versions declared on them, but we still allocate data
structures to contain these types anyway. This commit ensures we lazily allocate these containers in order to save
memory on every dependency that we declare.