Antlr plugin: Add antlrTool config to replace antlr config
This is to address #820, i.e. the fact that the antlr plugin
unnecessarily adds code generation tooling to the project's runtime
dependencies.

This adds new configurations antlrTool (for adding dependencies)
and antlrToolClasspath (for resolving those dependencies). For
backwards-compatibility purposes, these extend the antlr configuration.

Users using the antlr configuration -- or not declaring any antlr
dependency and using the default 2.7.7 dependency -- will have their
builds continue to work as before, but will receive warnings encouraging
them to move to the new approach.

As of this commit, I haven't confirmed with the Gradle team whether
I can/should add messaging around the antlr configuration being
removed in a future major version.

Signed-off-by: Alex Landau <alandau@cs.stanford.edu>