Deprecate Gradle.useLogger
The deprecation warning is emitted unconditionally, even when the same
call site triggers a configuration cache problem. This is done for
consistency and to highlight that the method is going to be removed
soon by linking the upgrade guide.

The implementation of useLogger delegates emitting the warning to the
DeprecatedFeaturesListener to keep all deprecation logic in one place.