Make CapabilitySelector subinterfaces private
There is still some question on how to best model this functionality.

Having a feature 'tack-on' to the module name of its owning component
can introduce ambiguity. For example, a module foo-bar with feature baz
and module foo with feature bar-baz have the same cabability when
using this model.

We may want to consider a fourth 'feature' field on a capability to alleviate
this. For now, we make the selector sub-interfaces internal to give us more
time to think about this