Remove ProjectDependency.getPath
Instead of exposing this, we remove upgrade recommendation to access mutable state of other projects.
That is bad advice and we do not recommend doing it at all.

Further, we refactor tests to not need to use the new method, by performing actual resolution and inspecting
the results instead of inspecting the dependencies before resolution