Enhance configuring of target bytecode version
When attempting to upgrade the toolchain to Java 21, we noticed some issues with the way we were configuring
java, groovy, and kotlin compilation.

1. For java, we did not use the release flag when compiling < 8, meaning projects targeting Java 6 could see java 8 APIs.
2. For groovy, we only set the source/target compatibility flags. When compiling with a newer toolchain, emitted Groovy
bytecode would contain classes from newer standard libs even though the code was targeted for older bytecode versions
3. For kotlin, we did not use the release flag, meaning when we used a newer toolchain, Kotlin would emit bytecode
that contained references to newer APIs that were not present on the target java version.

For most of these use cases, used the release flag more strictly (java and kotlin). Some projects actually use newer
APIs even though they target older versions. For this, we added an opt-in flag called 'usesFutureStdlib' that disables
the release flag.

For groovy, we are forced to use toolchains to set the proper version, as they do not support a --release flag. GROOVY-11105 tracks this feature.

Some projects also access JDK internals, which are not allowed when using the release flag. For this reason, we add a usesJdkInternals
to opt out of the release flag for certain projects.

Finally, some projects use dependencies that require a java version higher than the version they are targeting. We include a usesIncompatibleDependencies
for those use cases.

Each of these flags should be avoided if at all possible. Code owners of projects that use these flags should work to remove them.

Along the way, we had to make some minor adjustments to code, in order to comply with the more strict requirements given by the
use of the release flag.