Separate Signing and Security projects
The signing project contains the signing plugin. The security project contained both classes
for the signing plugin and for dependency management's dependency verification feature.

We move any classes from security that were only used for signing into the signing project.

Now, the security project no longer needs a core-api dependency.
The eventual goal is for dependency management to not need to depend on core or core-api