Fix type used in ResolutionFailureData cross-version test
The `org.gradle.api.problems.internal.GeneralData` type is an interface,
and Groovy was generating a dynamic proxy to allow viewing the actual
type, `org.gradle.tooling.events.problems.internal.GeneralData`, as an
instance of it when we cast with `as`. However, when running with a
different version than current as the TAPI provider, we do not put
Groovy on the classpath of the tooling types, so when Groovy went to
generate the proxy, it was unable to find Groovy types and failed. This
fixes the problem by using the correct type so a proxy does not need to
be created.