Use CC-aware representation of inherited environment variables
Exec/JavaExec tasks inherit environment variables of the current
process. Prior to this commit this was done naively, by snapshotting
them and storing everything in a simple Map in the task. This was
problematic because with CC, the snapshot was restored upon loading from
the cache, regardless of the changes to the build process' environment.

This commit introduces a separate storage of base environment
represented as a changing provider, and updates (changes, additions, and
removals). With that, the configuration can be replayed over the actual
environment.