Use contextual exception for non declarable error
We use an exception annotated with @Contextual so that the error appears in the console.

This brings up a bigger question whether this annotation is worth having or not, or if we should
just print all exceptions to the console. I can see there being many exceptions that are being hidden
because of this annotation and the logic behind it