Try configuring non-blocking on incoming/outgoing connector sooner
We were unconditionally setting blocking=false (non-blocking) just after binding
or connecting to the socket.

This was the source of some strange and hard to reproduce ioctl failures on
Windows. My suspicion is that this failure occurring so late gets treated
differently from a failure that occurs when connecting to the socket.

This change attempts to move configuring non-blocking on the socket to
an earlier point.

refs #30208