Allow ConfigurationCacheFingerprintWriter to process events concurrently
Normally, the ListenerManager holds a per-instance lock when dispatching
an event, so the listener implementation doesn't have to be thread-safe.
It also holds a per-listener interface lock for some internal reasons.
This, and the fact that the ConfigurationCacheFingerprintWriter can emit
events while processing events lead to deadlocks when two threads were
dispatching different kind of events to the writer.

This commit marks ConfigurationCacheFingerprintWriter as capable of
receiving events concurrently, thus eliminating the instance lock and
the potential for a deadlock.

The per-listener interface locks can in theory produce a deadlock too,
but in practice we can only aim for the second lock when processing
a ValueListener, so the potential for deadlocks is much lower and can be
addressed separately.