Add test cases for capability conflict between dependency and one of its transitives
We ran into a regression in the case where a dependency conflicts with one of its transtitives, where that transitive
is only referenced by that parent dependency.

This caused the transitive dependency to have no incoming edges at the time of conflict resolution. An attempted refactoring
to capability conflict resolution did not handle this situation correctly.

This introduces a number of tests intended to serve as a regression test against this scenario