Plugins can dynamically contribute additional build types to the init task
Additional project types can be added to a Build-scoped registry of available build types. If the registry is non-empty, it will be made available to init, replacing the built-in options. (There are now 2 separate branches of init task execution if contributed build types are found.)

Plugin<Setting>s can be dynamically loaded via a special system property org.gradle.buildinit.specs, so that init can be run in empty directories but have plugins applied that contribute specs.

Project generation is done entirely on the plugin-side, so this change is completely agnostic as to how projects are generated.

Infrastructure for parameterizing contributed project types is begun here, but left incomplete for now, as it is not vital to the first milestone of this work.