Reduce lock contention on local component provider
Reduce lock contention caused by trying to fetch project state each time a project component is requested.
The ProjectStateRegistry has an unconditional lock for each method call, which causes contention if called a lot.
We move the call to the registry inside the cache compute method to avoid calling the registry unless we need to.

We also make an improvement to the CalculatedValueCache implementation in CalculatedValueContainerFactory to
optimistically call get before calling computeIfAbsent, as computeIfAbsent will lock in some scenarios where get will not.
In cases where caches are expected to have a high hit rate and high parallelism, this avoids lock contention on the cache.