Create centralized cache factory
Create a centralized injectable cache factory. This cache factory implements optimizations
that avoid lock contention for read-heavy workloads in highly multithreaded environments.

This optimization is to call get first before calling computeIfAbsent, as computeIfAbsent
in some cases performs locking, where get does not.

Further, we expose caches that are useful in other scenarios, for example caches with
keys that are compared with identity, or caches whose values are calcualted within the
context of a CalcualtedValue, so they can acquire project locks in their loader function.

We migrate a number of existing caches to use this factory, and add a new cache to CachingAttributeSelectionSchema.java
which was determined as necessary based on profiling.

Another advantage of this pattern, which is not realized by this commit, is to allow centralized tracking
of caches. This could allow us to enable statistics capture and analysis for all caches in a build, allowing
us to determine if caches are performing well or if they are no longer necessary. This can be implemented in
future changes by leveraging the statistics provided by Caffeine caches.