Use jackson for build operation trace
We were previosuly using Groovy JSON for reading and writing JSON in the build operationt trace.
Groovy JSON only provides the ability to write to a string, not directly to a file. And, it cannot
pretty print JSON automatically.

That means, we need to write to a string, then pretty print, then write to a file. This non-inline
pretty printing is inefficient.

But, more importantly, we had to keep the entire JSON string in memory. This easily caused OOM errors
when running the build operation trace for large builds. Using Jackson, we can write directly to a file.
By writing directly to a file, we do not risk running out of memory when writing the final build operation
trace JSON file.

Without doing any performance analysis, this seems much faster than the Groovy JSON implementation.