Fix test failures
Use non-deprecated getNodeDependencies instead of the deprecated getDependencies property
on the build operation result, since we do not serialize deprecated properties

Jackson serializes byte arrays as base64 encoded instead of as an array of integers. This is a much
better serialization format for a byte array, so we update tests that originally expected a byte array
to decode the base64 encoded string

Add expectations for two more JARs in the distribution.

Fix some dependencies according to projectHealth