Use single progress shielding build operation listener
The ProgressShieldingBuildOperationListener is intended to ensure a progress event is not emitted
when its parent event is not in scope. It keeps track of each start and finish event for a given
build operation ID and only emits a progress event if the parent event is started but not finished

Previously, for each registered listener, we wrapped that listener in a progress shielding listener.
Now, we only create a single shielding listener and use that to guard all listners, as opposed to creating
multiple instances of the progress shielding listener.

This means we only keep track of the currently-in-scope build operations once, and not for each listener.