Some polish to testing
Update TreeFormatter.appendMethod to also print parameter types. Before, the output was ambiguous if there were overloaded methods
Update TestSuite.getTargets to return a DOC instead of a polymorphic DOC. This is a generic type and we should not require all implementations to have a polymorphic collection of targets
Do not implement Named for TestSuiteTarget. Targets should not need to be uniquely named, as we have discussed in the past, different identities like java version can be used to ID a target
Add javadoc to many test event listener classes, clarifying what they do with test events
Rename DefaultTestReport to HtmlTestReport, as it was the only TestReporter implementation
Delete the TestReporter interface, as it was only used to inject mocked implementations during testing. Update Tests to not use mocks
Add comments to AbstractTestTask executeTests method, explaining the steps of test execution and which test event listeners are registered