Allow a ResolutionBackedFileCollection to be created with a SelectedArtifactSet
Previsouly, it required a ResolutionReultProvider<SelectedArtifactSet>. A ResolutionResultProvider is able to provide
two values, one partial task dependency value and one full value. However, we only used the partial value when visiting
task dependencies and only used the full value when visiting artifacts.

Since SelectedArtifactSet itself can encode these two states, as it extends both TaskDependencyContainer and exposes
artifacts, we do not need to wrap it in a ResolutionResultProvider to create a FileCollection.

This opens up the possibility of creating file collections that are not based on configurations, for example those
used for artifact transform dependencies. These do not have the same partial/full values that configurations do,
and therefore need a way to construct file collections that are not wrapped in ResolutionResultProvider