Only pass strict resolver results to transform dependency resolver
We leverage the fact that after getTaskDependencyValue was called on the strict results, subsequent calls in the
same method would only call getTaskDependencyValue on the non strict results. The same goes for getValue. For this
reason, we can omit the non strict results at all, since if the call is premitted for the strict results, the
non strict results will always return the same value as the strict results did.

We also simplify a bit by removing the ResolutionResultProviderBackedSelectedArtifactSet wrapper. Since we eagerly
call getTaskDependencyValue and getValue, we can use the results from that directly in the FileCollection.

Finally for the approximate task dependency value, we skip wrapping in a ResolutionBackedFileCollection entirely,
since the SelectedArtifactSet we have is already a TaskDependencyContainer