Remove AbstractAttributesFactory and its redundant cache
The conversationCache added recently is quite expensive as it uses the entire map of attributes as the key.
In practice, it wraps an existing much more efficient cache that is already present in DefaultAttributesFactory.

Further, this cache was arbitrarily bounded to 1000 entries, making it easy for larger builds to go over the cache size
and render the cache inefficient. Since the cache uses the entire map as key, lookups can become very expensive, as the entire
map needs to be hashed on each lookup.

As a result, this can make the asImmutable call very expensive, contrary to the intention of the cache.

This commit reverts adding the cache and goes back to caching the result of asImmutable in the mutable attribute container.