Optimize injection of dependency management types with a large number of parameters
Injecting parameterized types into a constructor or method causes the ServiceRegistry and thus Java to reflectively parse
the method or constructor's generic signature, which is a String. This can take a non-trivial amount of time especially for
types with a large number of parameters.

We identify injection targets with long parameter lists and ensure they do not inject any generic parameters.