Back short circuit resolver with variant state
Resolution should be a function of only the root component and variant, plus any resolution strategy.
Having hasDependencies introduces multiple sources of truth that must be aligned.

We also remove markAsObserved and getAttributes, as markAsObserved happens automatically when observing the root component
and the attributes can be sourced from the root variant