Add coordination between reading shared objects and loading projects
It is possible for the shared object decoder thread to reach some
project-referring data faster than the main decoder thread sets up the
project structure for the build. It may happen when some build-level
data references shared objects and kicks off the auxiliary decoder.

The aux decoder now blocks if it encounters a project reference and
waits for the main thread to do its job.

Of course, the build is still going to be broken if the build-level data
has project references in it.