Apply review suggestions
Pull the explaination builder out of a static final field, as we don't know when the system property is set.
Rename LoadingCacheFactory to InMemoryCacheFactory, since we plan to also create non-loading caches in this factory.
Rename LoadingCache to InMemoryLoadingCache to reduce configuration with Guava's LoadingCache and caffeine's LoadingCache.
Use consistent parameter ordering for cache constructors.
Remove wrapper that prevents adding null values to the cache. ConcurrentHashMap already prevents this.