Detect edge case of resolving configuration removed from container
Detect and deprecate an edge case where a configuration with no dependencies is removed from the configuration
container and then resoled afterwards. We already fail in this case if the configuration has dependencies.
We previously short-circuited if the configuration did not have dependencies and did not fail in this case.

We skip documenting this deprecation since it is very unlikely for any user code to hit this and documenting it
would just lead to noise