Let `TypeDiscovery` use the real `PropertyExtractor` from the schema builder
This fixes an issue of `TypeDiscovery` not visiting types
mentioned in Java Bean properties. That happened because
the type discovery implementation made some assumptions about
the property extractor (used a specific implementation of it).

The API shape with `TypeDiscoveryServices` is chosen because
some more services might be added in the future when we start
handling certain generic types.