Remove ReusableAction
This was a workaround that we used for internal attribute disambiguation and compatibility rules as a performance optimization.

Since we knew our rules were stateless, we avoided instantiating them multiple times.

Third-party rules did not have this same benefit though.

Much improved caching in this area has now made it so that the overhead of instantiating rules is not significant.
We remove ReusableAction, making first-party and third-party attribute schema rules behave the same.