Don't do subscription clean up with the lock held
Cleaning up a subscription involves waiting for the worker thread to
finish processing. It is possible for the worker to try to acquire the
subscription lock again, leading to a deadlock.

Ideally, adding a listener after cleanup should throw, but this would be
a breaking change that is too big for a patch release.