Do not track isConstraint on selector
The constraint-ness of a selector can change throughout the graph traversal. Previously, we accounted for this by having a separate
selector for constraint and non-constraint selectors. However, this duplicated most of the selection logic between constraint
and non-constraint edges, even though the selection logic is the same.

This commit removes the constraint field from the selector and instead has each selector track the number of constraint and non-constraint
edges that use the selector. Then, it updates the module state accordingly as edges are used and released.