Rewrite method references in interfaces properly
The instrumentation expands references to interceptable methods into
full-bodied lambdas that calls the interceptor method. The lambda body
is held by a generated static "bridge" method, and a reference to the
latter replaces the original reference.

The generated handle of the bridge method wasn't taking into account the
type of the owner. However, when the owner of the bridge method is an
interface it should be represented in the method handle properly.
Otherwise, JVM fails to load the class.

This commit properly sets the type of the method handle.