Initial prototype for test suite auto reporting
Auto report on test targets by finalizing their test tasks by a report generator.
Only register the outputs of the target with the report if the target task is added to the
task graph, meaning we do not add extra inputs to the report generator task. This ensures
when the report runs, it only uses inputs from tasks that executed