Generate deprecated method when source is deprecated
Part of our decoration logic adds Closure taking methods for the Groovy
DSL when the Gradle API takes an Action.

These methods should be generated with a @Deprecated annotation when the
source method is itself deprecated.

While this does not have a direct user impact, if we ever publish
decorated DSL APIs, doing this would make them more correct.