Get rid of recursive descent in AbstractCollectionProperty calculation
Before this commit, a tree-like structure of PlusCollector emerged when
multiple entries were added with add/addAll. Getting anything out of
this structure required a recursive descent that wasn't even a tail
recursion. Sometimes, with many entries being preset, this caused a
StackOverflowError.

This commit replaces the tree with a flat structure inside the
CollectingSupplier, which can be traversed in O(1) stack space. The
suppliers are still effectively immutable, but `plus` causes a new
supplier to share the underlying list thus avoiding copying.

A consequence of the implementation change is that all absent-ignoring
operations are now handled by the CollectingSupplier itself.

Co-authored-by: Mikhail Lopatkin <mlopatkin@gradle.com>