Do not load file-system watching libraries on macOS 11
Previously we had file-system watching on macOS disabled (since Gradel 8.9),
but we still loaded the native libraries. This doesn't work anymore with
the new gradle-fileevents libraries we are including with Gradle 8.12,
and there is actually no point in loading the libraries if we are not
about to use them anyway.