Disable Classloader Caching in Isolated Workers
The classloaders for Isolated Workers are stored in a cache, but never releaed, and loaded. This results in every growing metaspace, which is crippling. There is no workaround. The only alternative is process isolation, which also requires additional memory and more system resources, which can also be unreasonably high.

In lieu of an actual fix this PR disables the caching for classpath isolated workers.

Workaround for #18313

Signed-off-by: Adam Semenenko <152864218+adam-enko@users.noreply.github.com>