Add additional context to configuration mutation deprecations
We added deprecations when a configuration that has been observed through resolution, consumption, or publishing -- of itself or a child configuration -- would become an error in Gradle 9.0

The deprecation is triggered when performing the mutation, but is 'enabled' when the configuration is observed. Since these happen at different times, the stack trace from the deprecation
does not provide any details on how to avoid the prior observation.

This commit attempts to capture some context during observation that can be exposed in the deprecation warning.

We can now tell, upon mutation, that a configuration was observed via publishing, resolution, or consumption as a variant.

One downside to this approach, is we cannot tell _which_ resolution consumed a given configuration as a variant.
This would be quite difficult to solve, or at least would require a lot of wiring that is not worth it.

For users who cannot resolve their deprecationw with this additional context, we should guide them to debugging Gradle to determine exactly how their configuration became observed