Intercept bound method references when the receiver is a subtype
Bytecode rewriter replaces method references to interceptable method
with references to generated static bridge methods. When the intercepted
method is an instance method, the bridge method accepts the receiver as
a first argument.

A bound method reference (e.g. `new File("..")::exists`) has special
rules for receiver type checking. It is allowed for the captured
receiver to be a subtype of the class that defines the method (e.g.
it can be `new MyFile("..")::exists`, though the bytecode still
references `File::exists`, unless, of course, the method is overloaded).
However, for the static method calls the types of captured variables and
the type of the referenced static method must match exactly.

Before this commit we were inferring the descriptor of the bridge method
based on the types of the target method. Of course, we were getting a
mismatch, because the actual receiver type was lost.

This commit adjusts the type of the receiver-argument of the generated
bridge method to match the declared type of the bound receiver to pass
the type check.