Do not realize all configurations in a project by default
Fixes #28192

Previously, unlike tasks, Configurations of a project are automatically realized, even if they are 'register'ed
or even if the new role-based factory methods are used, which are intended to be lazy.

This happens for two reasons
1. The base plugin had a configurations.all call in it in order to implement behavior related to the assemble task
2. When resolving a dependency against a target project, we would previously realize all configurations in the target project, even if they were non consumable.

This commit makes two significant changes
1. We no longer call configurations.all in the base plugin, but rather lazily add the artifacts of all visible configurations to the assemble task.
This prevents all configurations from being realized instantly if the base plugin is applied.

2. We no longer realize all configurations of a target project when resolving a dependency against it. We now only realize the configurations known to be
consumable or with an unknown/mutable role.

We add a new test to WellBehavedPluginTest to ensure it does not add configurations.all calls or other calls to instantly realize configurations
We also add a new test suite to verify dependency resolution only realizes consumable configurations or configurations with unknown an unknown role, but does not realize configurations with a known non-consumable type.