Automatically create test results elements for each test suite
Previously, we created a binary test results elements for each test target.
When performing aggregation, this only worked if the target test suite had a single target. Otherwise, we'd run into ambiguity.

To resolve this for multi-target test suites, we only create a single variant for each sutie and add all test results variants for all targets to the suite's variant.
This means when aggregating against a project, we will aggregate the results of all targets.

Eventually we will want to go back to a test results variant for each target, but until we have the targets able to declare their identity via attributes, we will have a single variant per suite.

We remove the concept of a test suite type, as this was a very simple mapping over the suite name, and only differed for the case of the default test suite, which has a test type of UNIT_TEST instead of its suite name.
We will eventually want to not use the suite name as an attribute either, but rather use the suite name as part of a capability.

The reports will eventually want to use attributes to distinguish which targets to aggregate instead of aggregating all targets from each suite.

We also make similar changes to Jacoco.