Back AttributeContainer with MapProperty
This will allow us to implement an addAll feature on AttributeContainers, allowing all
attribute from one container to be included in another container.

The implementation must be even more lazy than before, where before we know all attributes that
will be present eagerly, even if we don't know their values. Now, we do not know even the attribute
keys until the container is realized.

For this reason, validating actions and error checking some times may need to be delayed until the
container is realized.

Furthermore, this implementation does away with the immutable attributes cache. Previously, reading
the attribute values caused the container to realize all attributes and cache the result. Now, the container
will re-calculate the attributes each time the container is queried. This may lead to a performance degredation
that would require us to either optimize or to add-back caching and eager realization in some cases.