Pass ComponentSelector down resolver chain
The dependency resolver chain for external dependencies previously accepted an entire DependencyMetadata instance. This
includes details about the actual selector to resolve and the original requested selector before substitution. The resolver
chain does not need to know the original requested dependency.

Due to the prior structure, we needed to do some juggling between selectors and DependencyMetadata instances in cases
where we only had a selector but needed to query the resolver chain. This process was confusing and unnecessary.

Further, this opens the door to future changes to SelectorState. A SelectorState is an aggregate of a number of DependencyMetadata
instances and is responsible for resolving a components selector into a concrete component identifier. When a SelectorState
incorporates multiple dependnecy metadata instances, it only passes the first dependency metadata instance when executing
the resolver chian.

This commit updates the resolver chain to accept the raw selector only, in addition to a ComponentOverrideMetadat. In future changes,
we can update the ComponentOverrideMetadata to incorporate details from all aggregated DependencyMetadata instances instead of just the
first instance