Refactor LenientPlatformGraphResolveState to not extend DefaultExternalComponentGraphResolveState
A lenient platform component is unlike any other external components.

It does not have configurations or variants in the conventional sense, so prior logic did some odd things in order to make
a lenient platform component conform to the format required for an external component state.

This commit reimplements lenient platform components standalone, to avoid some confusing logic and untangle external and module components
so that we can  make further changes to module components in future changes