Avoid HasAttributes in internal attribute matching
AttributeMatcher and MultipleCandidateMatcher relied on candidates implementing HasAttributes.
There are two main problems with this interface:

1. It is public, so it only exposes AttributeContainer instead of ImmutableAttributes. Our attribute
matching code needed a number of casts and copies in order to deal with this. This introduces clutter,
but also seems to break some optimizations in the JVM due to the casting and extra method calls, as
methods like DefaultAttributeMatcher.CachedQuery.from have come up as hot spots in flame graphs.
2. It hurts IDE navigability, as ctrl+clicking on any type that implements HasAttributes shows _all_ usages
of the interface across the entire build, even if they are not related. Having such a wide-ranging interface
like this almost always results in poor IDE experience. We should deprecate and remove HasAttributes in public
types as well, but this is a start.

Since ImmutableAttributes does not implement AttributeMatchingCandidate directly, we also introduce ImmutableAttributesBackedMatchingCandidate.
This wrapper makes the attribute matching done in publishing a bit more clumsy. But, this seems like a worthy trade-off