Resolve bug in dependency resolution
When we restart an edge due to capability conflict or module version selection change,
we may encounter an edge where the source node is not selected.

Previously, we would skip restarting (retargeting) this node in that case, but it may be possible that
the source node re-enters the graph. In this case, Gradle expects the node and its children to still be
vaild subgraphs, unless it otherwise deconstructed that subgraph.

So, we update the edge-restart code to maintain the state of these detached subgraphs when edges are re-targeted.