Produce JUnit XML timestamps that include milliseconds
ISO_INSTANT is the JDK built-in way of representing a Instant. An Instant
is an instantaneous point on the time-line. The start time of a test is an
instant.

Gradle's timestamps were always UTC-based, but we did not include the time
zone information. ISO_INSTANT is explicitly UTC (Z).

Parsers, like Jenkins, can handle parsing a JUnit XML with timezone information.