Use ImmutableSet when caching task graph node values
This cache gets retained in heap for the entirety of cache graph walking.
For very large graphs, this cache can retain tens of GB of memory.
The values of this cache are immutable, yet we store a mutable set within it -- a LinkedHashSet -- which is very memory inefficient

In the Android 6k project benchmark project, when running assemble, this cache was found to contain over 124 million LinkedHashMap entries,
totaling almost 8GB of retained heap _for the map nodes themselves_, not including the contents of the map.

This commit stores an immutable Set instead, ensuring we use a much more space efficient data structure for this cache.