Use a non-dashed name for Kotlin script-compiler-environment compiler arg
We use to use `implicit-import` as a key in script compiler environment to convey external plugin spec builders. But now using a KotlinCompile task instead of the embedde compiler triggers some parsing/validation of the script-compiler-environment string that chokes on dashes in key names.

This commit changes the key name to be camel-cased and introduce a constants to make it easier to find usages.