Make MutationInfo immutable
MutationInfo, with the exception of the nodesYetToConsumeOutput state, was generated wholly by the ResolveMutationsNode.
After the ResolveMutationsNode executes, the MutationInfo is not mutated further.
For very large graphs, this MutationInfo can retain a significant amount of memory -- in part due to memory inefficient mutable
data structures used to store its state.

This commit ensures we calculate the entirety of MutationInfo at once, so that we can store its state in a memory efficient immutable
data structure. This greately reduces the retained heap from this state.