Rename requireFeature to requireSuffix
A feature name is traditionally camelCase. This is an ecosystem level abstraction.

We want to have a core dependency management functionality allowing users to append _some_ suffix to a module name
to get the capability name, but we don't want to arbitrarily add string manipulation into the mix in this core DM functionality.

When requireFeature was introduced, it did not have string manipulation in it. However, this confused users who were used to
having their string converted to kebob case for them.

In this commit, we generalize requireFeature to what it actually is, a suffix. We will continue to reserve the name 'feature' for a
camel case string, which when converted to kebob case and prefixed with a dash, is the proper suffix to append to a module name.

Future changes will add a feature dependency modifier